Add validators for duration and category
